Improved tests and improved implementation of the Warehouse class

+++ src/Warehouse/Warehouse.h
@@ -43,7 +43,7 @@ public:
      * @param newCapacity The new capacity of the warehouse.
      * @return status SUCCESS if the operation is successful, ERROR otherwise.
      */
-    status updateStatus(double newCapacity);
+    status updateStatus(int newCapacity);
 
     // Product operations
